ScriptCommand - Fix relative path handling. Add debug output.
The `php:script` subcommand would fail to run scripts in the current
directory -- because it attempted to load them post-boot (in the CMS's root
directory).

With this revision, we default to checking in the original CWD.  For
backward compat, we'll also fallback to the CMS CWD if necessary and display a
warning.