ext:download - Detect existing extensions. Add -k/-f options.
When you run `cv dl foo bar`, and `foo` already exists, it's unclear
what the best course of action is -- there are use-cases for
(a) updating to the latest `foo` or (b) keeping `foo`.

With this revision, the default behavior is to ask the user what to do.  But
for scripting purposes, you can use `-f` (force download) or `-k` (keep).