Bootstrap - Fix for Joomla booting underneath `administrator`
The bootstrap searches for the Joomla root by locating a key-file, one of
which can be `components/com_users/users.php`.

It turns out there are actually two similarly named files:

 * `components/com_users/users.php`
 * `administrator/components/com_users/users.php`

The problem arises if the PWD is something like this:

 * `/srv/buildkit/build/hydra-joomla/administrator/components/com_civicrm/`

The directory search finds a false-positive when checking if
`/srv/buildkit/build/hydra-joomla/administrator` is the Joomla root.

This changes the search to use a different key-file which is not as likely
to have a false-positive.