ApiCommand - Make the verbose output useful. Display backtraces more easily.
APIv3 accepts a `debug` option and returns a `trace` field (if there's an
exception).  In the normal output formats, the `trace` is very hard to read.
Alas, that is necessary to ensure that output complies with the requested
output formats (e.g.  JSON).

However, if the caller has requested verbose mode (-v), then it seems fair
to dump more info on the screen without concern for the requested format.
Thus, in verbose mode, we can print the backtrace in a nicer way.