(POC) Update symfony/console 2.x => 3.x
This patch updates symfony/console.  A straight update, however, causes a
regression: if you run `php:eval` or `php:script` with an invalid script,
it will no longer have a semantically appropriate error code. Also,
the formatting of the error seems less appropriate.

This patch addresses the exit code (which has more functional impact on
scripting and unit-tests), but I'm kind of surprised that `symfony/console`
went from (v2) reporting the error well to (v3) not reporting the error as
well.  It makes me think that maybe they changed the design and have an
alternative in mind.