"url"/"open" - Add option to auto-login
Background
----------

The "url" command allows you to generate a URL pointing to Civi, eg

```
$ cv url 'civicrm/foo?reset=1'
"https://example.com/civicrm/foo?reset=1"
```

The "open" command is an alias.  On Linux and OSX workstations, it will
generate the URL and open it in the web-browser.

```
$ cv open 'civicrm/foo?reset=1'
```

Changes
-------

This adds an option `cv url --login` (aka `cv open --login`) which appends
an authentication-token to URL.  For example:

```
$ cv open 'civicrm/foo?reset=1' --user=demo --login
$ cv open 'civicrm/foo?reset=1' --user=admin --login
```