UrlCommandTrait - Take a more neutral default on "frontend vs backend"
The contract for `CRM_Utils_System::url()` is a bit confusing -- it will generate
URL in the frontend XOR backend, but it accepts two arguments (`bool $frontend`
and `bool $forceBackend`), which both default to `FALSE`.

TBH, I don't entirely understand the intended usage for these two options.

The prior behavior of `cv url` was to force on or the other to be true. This
made it explicit, but it meant that it never did the *default thing*.

In some testing with `cv http`, I'm finding that the *default thing*
(whatever that means) seems to be more useful.