Let admin choose default boot proto. Respect both CIVIRM_BOOT and CIVICRM_SETTINGS.
This is preparation for v0.4. The ultimate aim is to flip the default
boot behavior (`full/Bootstrap` <=> `cms-full/CmsBootstrap`); but this
is a big change. This patch is about allowing admins fine-tune
the transition.

* In v0.3, the default is `full/Bootstrap`.
* In v0.4, it will be `cms-full/CmsBootstrap`.
* If you're on v0.3 and want to get ahead (*start using `CmsBootstrap`
  by default; now*), then you can explicitly set `CIVICRM_BOOT=Auto://`
* If you're on v0.4 and want to stay behind (*stay on `Bootstrap` by
  default*), then you can explicitly set `CIVICRM_SETTING=Auto`

These env-vars already exist, but they're not consistently respected.  You
sometimes need to manually set `--level=full` or `--level=cms-full`.  This
patch allows you switch without needing to manually pass `--level=X` with
each call.