FillCommand - Fix failing test
This patch strictly does fix a bug -- when you use the `--file` option, it will
fail to read properties like `CIVICRM_SETTINGS_PATH`. This smells like a
bad thing, and it raises trouble in php81.

However, the `--file` option should be quite obscure - iirc, it mostly
existrs to support the test-case.