Drupal 9 - Fix error regarding `AccountProxy::get()` (#148)
This should fix an issue that appears in both #148 and in the test-matrix. It presents as:

```
Error: Call to undefined method Drupal\Core\Session\AccountProxy::get() in
CRM_Utils_System_Drupal8->getUserIDFromUserObject() (line 595 of
/home/jenkins/bknix-max/build/build-1/vendor/civicrm/civicrm-core/CRM/Utils/System/Drupal8.php).
```

And it flows from a call to `\CRM_Core_BAO_UFMatch::synchronize`. There are, in fact, two
callers for `synchronize()`, and I believe both need updating.