Declare "psr/log" requirement. Fix "AbstractLogger" error with "core:install"
Fixes a recent regression where "core:install" outputs this:

  Found code for civicrm-core in /srv/runner/home/runner-1/buildkit/build/build-1/web/sites/all/modules/civicrm
  Found code for civicrm-setup in /srv/runner/home/runner-1/buildkit/build/build-1/web/sites/all/modules/civicrm/setup
  Error: Class &#039;Cvphar\Psr\Log\AbstractLogger&#039; not found in include() (line 10 of phar:///srv/runner/home/runner-1/buildkit/bin/cv/vendor/composer/../symfony/console/Logger/ConsoleLogger.php).

In non-prefixed environments, this worked because `psr/log` was generally provided by some other means.

But now that cv.phar prefixing has been restored, it's looking for an isolated version of "AbstractLogger".
That seems good in terms of avoiding future conflicts. The problem is that "AbstractLogger" actually needs
to be available to "cv".