Define `Top` helper to load classes regardless of namespace prefixing
Currently, `cv` has a list of php-scoper rules to exclude certain classes from prefixing.
The same list has to be adapted to use in `civix` - so, presumbly, they'll have to be
sync'd over time.

The main reason for this is to allow the PHAR to call-out to the UF during bootstrap -- and
each UF has a different list of rules.

In theory, we could drop these rules -- and replace any lines like
`drupal_bootstrap()` with `Top::call('drupal_bootstrap')`. Then there would be no
need to keep the rules in sync.

Additionally, it would allow `cv.phar` to take advantage of libraries (like `symfony/event`)
that might be dual-purpose. (Ex: cv.phar and UF both have `EventDispatcher`s; and you want
to add listeners to both of them.)