cv flush - Remove extra call to boot()
As part of the reorg to use CvCommand, most of the explicit calls
to `boot()` were dropped (ae1ed440417b93076aa1e34b7fdc1f4978b33188)
in favor of the inherited/automatic boot.

This one fell through the cracks.  The old call to `$this->boot()` probably
should've been dropped at the same time (or else FlushCommand should've been
flagged as manual-boot).

In any event, this call is duplicative -- and in some configurations, it
leads to a chain of other errors (#220).