ErrorHandler - Present more error info about MySQL connections (work-around)
Overview
--------

Present more error info about MySQL connections. This may be most useful if you're debugging a failure
in MySQL-SSL.

Technical Details
-----------------

This is a work-around for a quirk in `DB_mysqli::connect()` -- it suppresses warnings
generated by `mysqli_real_connect()`. But those can be useful!

Th work-around depends on your boot mode (Civi-first vs CMS-first).
It applies to Civi-first. For CMS-first, the CMS error-handler is active
at the critical moment -- and it doesn't have such a work-around.

I'm on the fence about merging this. This is safe enough (very narrow
impact). But it would be cleaner (and probably better) to patch
`DB_mysqli::connect()` to simply allow warnings. (Or, failing that, buffer
the warnings -- and let them through if we hit any subsequent problems.)
Addressing that root cause would make it less sensitive to boot-mode.

Of course, warnings have been suppressed for so long... it's hard to know if
we might create some backlash from it...