CmsBootstrap - Fix excessively modest error-reporting
Overview
--------

When you call out the CMS to bootstrap the environment (CMS-first protocol),
then it may sometimes adjust the PHP `error_reporting()` policy (eg to hide
warnings from anonymous users). It's well and good for CMS to dictate that
policy on web UI. But for CLI, we often prefer to communicate more to the
console user.

With this, the effective value of `error_reporting()` should be more consistent
between traditional `Bootstrap.php` (Civi-first) and `CmsBootstrap.php` (CMS-first).

Before
------

Once `CmsBootstrap` has started the CMS, it tries to reset the value.
But the specific value used here is a poor choice. (I suspect there was
an inversion error in the original draft -- `E_ALL` vs `1`. Either value is
suboptimal, but `1` is particularly bad.)

After
-----

Once `CmsBootstrap` has started the CMS, it restores the prior value.
This means that we preserve whatever value was defined by `php.ini`
or `bin/cv` or `php -d error_reporting=` or similar.